requirements.txt
autopep8==1.4.4
awscli==1.16.283
beautifulsoup4==4.8.1
boto==2.49.0
boto3==1.10.19
botocore==1.13.19
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
Django==2.2.5
django-appconf==1.0.3
django-bootstrap4==1.0.1
django-cleanup==4.0.0
django-environ==0.4.5
django-imagekit==4.0.2
django-sendgrid-v5==0.8.0
django-storages==1.7.2
docutils==0.15.2
entrypoints==0.3
factory-boy==2.12.0
Faker==2.0.3
flake8==3.7.8
future==0.18.2
idna==2.8
jmespath==0.9.4
mccabe==0.6.1
pilkit==2.0
Pillow==6.1.0
pyasn1==0.4.8
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.0
python-http-client==3.2.1
pytz==2019.2
PyYAML==5.1.2
psycopg2==2.8.4
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
sendgrid==6.1.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
stripe==2.37.2
text-unidecode==1.3
urllib3==1.25.6
gunicorn==19.9.0 
